+++ src/main/java/baekgwa/suhoserver/domain/project/dto/ProjectResponse.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 import baekgwa.suhoserver.domain.project.type.ProjectSort;
 import baekgwa.suhoserver.model.branch.bom.entity.BranchBomEntity;
-import baekgwa.suhoserver.model.project.branch.entity.ProjectBranchEntity;
+import baekgwa.suhoserver.model.project.branch.branch.entity.ProjectBranchEntity;
 import baekgwa.suhoserver.model.project.project.entity.ProjectEntity;
 import baekgwa.suhoserver.model.project.straight.straight.entity.ProjectStraightEntity;
 import baekgwa.suhoserver.model.straight.info.entity.StraightInfoEntity;
