+++ src/main/java/baekgwa/suhoserver/domain/project/service/ProjectBomService.java
@@ -35,8 +35,8 @@ import baekgwa.suhoserver.infra.excel.util.ExcelStyler;
 import baekgwa.suhoserver.model.branch.bom.entity.BranchBomEntity;
 import baekgwa.suhoserver.model.branch.bom.repository.BranchBomRepository;
 import baekgwa.suhoserver.model.branch.type.entity.BranchTypeEntity;
-import baekgwa.suhoserver.model.project.branch.entity.ProjectBranchEntity;
-import baekgwa.suhoserver.model.project.branch.repository.ProjectBranchRepository;
+import baekgwa.suhoserver.model.project.branch.branch.entity.ProjectBranchEntity;
+import baekgwa.suhoserver.model.project.branch.branch.repository.ProjectBranchRepository;
 import baekgwa.suhoserver.model.project.project.entity.ProjectEntity;
 import baekgwa.suhoserver.model.project.straight.straight.entity.ProjectStraightEntity;
 import baekgwa.suhoserver.model.project.straight.straight.repository.ProjectStraightRepository;
