+++ src/main/java/baekgwa/suhoserver/domain/project/service/ProjectWriteService.java
@@ -16,8 +16,8 @@ import baekgwa.suhoserver.domain.project.dto.ProjectRequest;
 import baekgwa.suhoserver.global.exception.GlobalException;
 import baekgwa.suhoserver.global.response.ErrorCode;
 import baekgwa.suhoserver.model.branch.type.entity.BranchTypeEntity;
-import baekgwa.suhoserver.model.project.branch.entity.ProjectBranchEntity;
-import baekgwa.suhoserver.model.project.branch.repository.ProjectBranchRepository;
+import baekgwa.suhoserver.model.project.branch.branch.entity.ProjectBranchEntity;
+import baekgwa.suhoserver.model.project.branch.branch.repository.ProjectBranchRepository;
 import baekgwa.suhoserver.model.project.project.entity.ProjectEntity;
 import baekgwa.suhoserver.model.project.project.repository.ProjectRepository;
 import baekgwa.suhoserver.model.project.straight.serial.repository.ProjectStraightSerialRepository;
